Monkey ROL
This technique, which was modified from previously published methods (A. L. Daitch et al., 2016; B. L. Foster et al., 2015), estimates the time of the inflection point of a task-related activation or deactivation. Since it is difficult to estimate onsets of single-trial time courses (given lower SNR), we estimated a distribution of measures for each electrode site using a bootstrap resampling procedure (repeated 1000 times), in which we selected a subset of trials in each repetition, and approximated the onset based on the average time course across this subset of trials. For each bootstrap sample, we divided the post-stimulus signal (0-1000 ms) into 60ms bins with 58ms overlap, and the data within each bin was averaged and fitted linearly to estimate slope. Next, a threshold was defined as the mean plus (or minus for deactivation) one SD of the baseline distribution to isolate signal fluctuations above the baseline period (i.e. 200 ms ITI prior to stimulus onset across all trials in the bootstrap repetition), and we identified the earliest fifty consecutive bins in which the average power amplitude exceeded this threshold.  
From this set, we noted the bin with the steepest slope and averaged this slope with the overall slope of the baseline period to define a threshold. In a stepwise manner, backwards from the window containing the steepest slope, we marked the first bin to show a slope smaller than the slope threshold (for activations) or larger than the slope threshold (for deactivations). The earliest time-point of this bin is marked as the signal onset in a bootstrap iteration. We defined the onset for a site as the median value across the bootstrap estimates. This analysis was implemented on non-smoothed HFB signals to minimize confounds associated with temporal smoothing.